postinst: fix whiptail parameters
This fixes whiptail parameters. There was an extra parameter which caused an error. Converted to use title instead.
Also, since we are in errexit mode, we should be careful that calls that can fail be catched, otherwise the whole apt process will fail and leave the system in a broken and potentially unbootable state. Updated i2cget calls to output sane values even if an error occurs when trying to access the i2c bus. This happened to me when trying to update on an older kernel. A warning message might be good here also to warn users to reboot into a recent kernel before performing the update.